Added API endpoint to send text email to a list of recipients
