+++ .env.example
@@ -10,3 +10,5 @@ CYBERSOURCE_TRANSACTION_KEY=
 CYBERSOURCE_SECURE_ACCEPTANCE_URL=
 CYBERSOURCE_PROFILE_ID=
 CYBERSOURCE_REFERENCE_PREFIX=
+MAILGUN_KEY=
+MAILGUN_RECIPIENT_OVERRIDE=
