+++ search/views.py
@@ -4,20 +4,16 @@ Views for the Search app
 
 import logging
 
-from django.conf import settings
-from elasticsearch_dsl import Search, Q
 from rest_framework import status
 from rest_framework.views import APIView
 from rest_framework.authentication import SessionAuthentication
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 
-from roles.api import get_advance_searchable_programs
-from search.api import (
-    DOC_TYPES,
-    get_conn
-)
 from search.permissions import UserCanSearchPermission
+from search.api import prepare_and_execute_search
+from search.exceptions import NoProgramAccessException
+
 
 log = logging.getLogger(__name__)
 
@@ -29,40 +25,17 @@ class ElasticProxyView(APIView):
     authentication_classes = (SessionAuthentication, )
     permission_classes = (IsAuthenticated, UserCanSearchPermission, )
 
-    def _search_elasticsearch(self, request):  # pylint: disable=no-self-use
+    def _execute_search_from_request(self, request):  # pylint: disable=no-self-use
         """
         Common function that will take care of handling requests coming from different methods.
         """
-        # make sure there is a live connection
-        get_conn()
-
-        # create a search object and load the query coming from the client
-        search = Search(index=settings.ELASTICSEARCH_INDEX, doc_type=DOC_TYPES)
-        search = search.from_dict(request.data)
-
-        # extract all the programs where the user is allowed to search
-        users_allowed_programs = get_advance_searchable_programs(request.user)
-        # if the user cannot search any program, return an error
-        # in theory this should never happen because `UserCanSearchPermission`
-        # takes care of doing the same check, but better to keep it to avoid
-        # that a theoretical bug exposes all the data in the index
-        if not users_allowed_programs:
+        try:
+            results = prepare_and_execute_search(request.user, search_param_dict=request.data)
+        except NoProgramAccessException:
             return Response(
                 status=status.HTTP_403_FORBIDDEN,
                 data={'error': 'no_available_programs'}
             )
-        # no matter what the query is, limit the programs to the allowed ones
-        # if this is a superset of what searchkit sends, this will not impact the result
-        query_limit = Q(
-            'bool',
-            should=[
-                Q('term', **{'program.id': program.id}) for program in users_allowed_programs
-            ]
-        )
-        search = search.query(query_limit)
-        # execute the query
-        results = search.execute()
-        # return the result as a dictionary
         return Response(
             results.to_dict()
         )
@@ -71,10 +44,10 @@ class ElasticProxyView(APIView):
         """
         Handler for GET requests
         """
-        return self._search_elasticsearch(request)
+        return self._execute_search_from_request(request)
 
     def post(self, request, *args, **kwargs):
         """
         Handler for POST requests
         """
-        return self._search_elasticsearch(request)
+        return self._execute_search_from_request(request)
